remove models package
