+++ app/handlers/settings_test.go
@@ -9,12 +9,11 @@ import (
 
 	"github.com/getfider/fider/app"
 
+	"github.com/getfider/fider/app/models/entities"
 	"github.com/getfider/fider/app/models/enum"
 	"github.com/getfider/fider/app/models/query"
 
 	"github.com/getfider/fider/app/models/cmd"
-
-	"github.com/getfider/fider/app/models"
 
 	"github.com/getfider/fider/app/handlers"
 	. "github.com/getfider/fider/app/pkg/assert"
@@ -222,7 +221,7 @@ func TestVerifyChangeEmailKeyHandler_Success(t *testing.T) {
 	key := "th3-s3cr3t"
 	bus.AddHandler(func(ctx context.Context, q *query.GetVerificationByKey) error {
 		if q.Key == key && q.Kind == enum.EmailVerificationKindChangeEmail {
-			q.Result = &models.EmailVerification{
+			q.Result = &entities.EmailVerification{
 				UserID:    mock.JonSnow.ID,
 				Key:       q.Key,
 				Kind:      q.Kind,
@@ -266,7 +265,7 @@ func TestVerifyChangeEmailKeyHandler_DifferentUser(t *testing.T) {
 	key := "th3-s3cr3t"
 	bus.AddHandler(func(ctx context.Context, q *query.GetVerificationByKey) error {
 		if q.Key == key && q.Kind == enum.EmailVerificationKindChangeEmail {
-			q.Result = &models.EmailVerification{
+			q.Result = &entities.EmailVerification{
 				Key:       q.Key,
 				Kind:      q.Kind,
 				ExpiresAt: time.Now().Add(10 * time.Minute),
