+++ app/tasks/tasks.go
@@ -9,7 +9,6 @@ import (
 
 	"github.com/getfider/fider/app"
 	"github.com/getfider/fider/app/actions"
-	"github.com/getfider/fider/app/models"
 	"github.com/getfider/fider/app/models/cmd"
 	"github.com/getfider/fider/app/models/dto"
 	"github.com/getfider/fider/app/models/entities"
@@ -353,7 +352,7 @@ func NotifyAboutDeletedPost(post *entities.Post) worker.Task {
 }
 
 //SendInvites sends one email to each invited recipient
-func SendInvites(subject, message string, invitations []*models.UserInvitation) worker.Task {
+func SendInvites(subject, message string, invitations []*dto.UserInvitation) worker.Task {
 	return describe("Send invites", func(c *worker.Context) error {
 		to := make([]dto.Recipient, len(invitations))
 		for i, invite := range invitations {
